Robustness fix (#16)
@m-cespa small changes. Runs. Will take ages at current due to massive
AmpleGCG preprocessing overheard. Great stuff.

---------

+++ experiments/uncertainty/engine.py
@@ -65,7 +65,7 @@ def train(model, exp_config: ExperimentConfig, optimizer, scheduler):
             loss_logits = get_binary_logits(logits, t_config)
 
             loss = torch.binary_cross_entropy_with_logits(
-                loss_logits, labels
+                loss_logits, labels.float()
             ).mean()
 
             loss.backward()
